Refactored User and Course Controllers
Moved functionality to the corresponding User and Course controllers
